+++ src/components/icons-and-geometry/index.ts
@@ -1,0 +1,5 @@
+export * from './Avatar'
+export * from './Circle'
+export * from './HelpwaveLogo'
+export * from './Ring'
+export * from './Tag'