+++ src/components/properties/index.ts
@@ -1,0 +1,7 @@
+export * from './CheckboxProperty'
+export * from './DateProperty'
+export * from './MultiSelectProperty'
+export * from './NumberProperty'
+export * from './PropertyBase'
+export * from './SelectProperty'
+export * from './TextProperty'