+++ src/components/table/index.ts
@@ -1,0 +1,6 @@
+export * from './FillerRowElement'
+export * from './Filter'
+export * from './Table'
+export * from './TableCell'
+export * from './TableFilterButton'
+export * from './TableSortButton'