+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.27] - 2025-09-29
+## [0.1.28] - 2025-10-02
+
+### Added
+- added barreling script and barrel files
+
+### Changed
+- pin dependencies
+
+### Fixed
+- Fix css export path
+
+## [0.1.27] - 2025-10-02
 
 ### Added
 - Add a theme preference listener to `useTheme` hook